Mor Cohen

+++ FVCalc.java
@@ -3,7 +3,7 @@ public class FVCalc {
 	public static void main(String[] args){
 		// user enters Current value, Rate, Number of years
 		int currentValue = Integer.parseInt(args[0]);
-		int rate = Integer.parseInt(args[1]);
+		double rate = Double.parseDouble(args[1]);
 		int n = Integer.parseInt(args[2]);
 		// creating a new double value which stands for the future value after the rate increase and using the given formula
 		// [Current Value * (1 + Rate)^Number of years]
